[UP] move optim step funcs
